Update clr_bundle_add command to install containers-basic
From Clear Linux* release 5050 the instruction to install a
new bundle has changed from clr_bundle_add to swupd bundle-add.

This change updates the instructions to install the
containers-basic bundle to be compatible with the most recent
Clear Linux* image released in public (5140 for native/cloud/kvm
and 5210 for container image).