ciao: Update ciao-launcher instructions
This commit removes the --network option that was specified on the various
ciao-launcher example command lines in the previous version of this
document.  --network is no longer needed as the role of a given launcher
instance is now determined from its SSNTP certificate.

The instructions for resetting compute and network nodes have been updated.
The --hard-reset option for launcher makes the other steps unnecessary.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>