ciao: freshen documentation on certificate authorities
We removed a section on adding certificates to your browser, yet we do need
info on adding a certificate to the system CA trust.  But this varies with
each linux distribution.  It's complicated and authoritative information on
the topic is out there in web, so give minimal information to help a
reasonable system administrator on their way.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>