getting_started: Update broken 'Clear Container' text to 'Docker image'
The 'Container Evaluation' section of the 'Getting Started' page was
just wrong - it referenced the image files that are used as part of the
Clear Containers execution system, and implied you could execute them
directly, which you cannot in any useful way.

As this page is really about Clear Linux, and we now do actually have
Clear Linux images on the main Docker hub, update the whole section to
describe how you can run Clear Linux under Docker, and leave the Clear
Containers details to the existing Clear Containers pages.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>