Instructions for NATing cluster on network boot
The description of how to construct your network was missing from the
network_boot documentation. This adds instructions on how to
configure the cluster so that the PXE / DHCP server allows traffic
to flow from internal to external networks and vice versa. Without
this the documentation is incomplete as it assumes the user knows how
to set up some kind of networking. By providing this information the
user is able to get a fully functioning cluster up an running biased
solely on this document.

Signed-off-by: John Andersen <john.s.andersen@intel.com>