Modify file structure of docs source files.
Changed the file structure to allow for easier and better maintenance of
the docs. Files were renamed as needed and following a consistent
convention. Added a index.rst to every folder allowing for partial
builds of the docs. Included content previously left out of the docs
build. The commit also performs minimal housekeeping.

Signed-off-by: Rodrigo Caballero <rodrigo.caballero.abraham@intel.com>