ciao: yet another many months out of date command line invocation
I would swear I searched for and fixed these all before, but clearly
this delete command is broken and should have been fixed in the past.
Any invocation that is, eg:

ciao-cli -something

is a sign of a broken command, as all commands now start with, eg:

ciao-cli something

Later option arguments may include a dash, but none of the top level
commands/subcommands do.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>