Edit grammar, spelling, sentence structure, tone, style, and trademarks.
Performed an edit ensuring parallel content structures, a consistent
tone and style, good grammar and spelling, and trademark
references only on the first instance. Added the name of the bundle at
the begining of each paragraph in the exact form users need to put in
the commands, thus providing a better user experience. Lastly, I fixed
capitalization issues with the links. Although the GitHub repos are
named in lower case, that does not mean the links' text  should be in
lower case too.

Signed-off-by: Rodrigo Caballero <rodrigo.caballero.abraham@intel.com>