Restructure all technical content.
This restructure organizes the content into three sections, one per
technology. The content was left as untouched as possible, however
several directives needed to be standardized in order to implement the
new structure. All filenames follow use dashes-as-spaces and are written
using lower-case. Standard labels for files were implemented using the
filename as the label, thus, content can be cross-referenced without
having to open the file to see the label. Figures were moved to
decentralized folders to enable modularity. The scope of the restructure
as well as the dependencies between the changes required me to do it in
a single commit. Lastly, some cleanup had to be done for the
standardized structure to function properly.

Signed-off-by: Rodrigo Caballero <rodrigo.caballero.abraham@intel.com>