conf.py: fix html_add_permalinks
When compiling documentation a warning message is showed:

WARNING: The config value `html_add_permalinks' has type `bool', defaults to `str'.

The documentation from http://www.sphinx-doc.org/es/stable/config.html
said to set it to None or the empty string to disable permalinks.

Now it is set to empty string and the warning disappear.