Correct update content validation commands
To verify the Manifest.MoM, the signature of the Manifest.MoM should be compared with the Swupd certificate, not the Clear Linux certificate.  Also, adding a section for manually validating the Swupd certificate so that it matches the section for manually validating the Clear Linux certificate.  Making the distinction between the Clear Linux certificate and the Swupd certificate more clear.