Update Mixer documentation
Updating the documentation for mixer to account for recent changes,
and to better clarify some concepts. This is to coincide with an
upcoming release of mixer, in which several important changes will
be introduced.

1) The mixer CLI has been rewritten, supporting POSIX-style flags
and a modern command layout. This is a breaking change, as commands
and flags have been renamed.

2) The initialization steps for mixer have been further automated,
so the documentation has been updated to account for the new,
simpler approach.

3) Documentation has been added for the recent command for
automating the addition of bundles to a mix.

4) The tool now handles caching upstream CLR bundle definitions
in a more automated way, and so the 'mixer get-bundles' command
(and its corresponding documentation section) has been removed.

Signed-off-by: Kevin C. Wells <kevin.c.wells@intel.com>