Change curly quotes for straight quotes
In the "Change the Default Configuration and Data
Directory" section, the curly quotation marks in the
httpd.conf example cause service start errors when
the text is copied and pasted over directly to the new
httpd.conf file, as part of following the section's
steps.

This commit replaces the curly double quotes with straight
double quotes to correct the wrong behavior.

Signed-off-by: Jair Gonzalez <jair.de.jesus.gonzalez.plascencia@intel.com>