Correct the default data directory references
The default data directory for Apache is /var/www/html
in the latest version (tested on VERSION_ID=20490), instead
of /usr/share/httpd/htdocs as specified in the web-server-install
tutorial. This causes several inconsistencies in the "Change the
Default Configuration and Data Directory" section.

This commit updates the steps to configure a different data
directory for demostrative purposes (/var/www/tutorial) as the
proposed one (/var/www/html) is currently the default value.
As the rest of the document and the WordPress server tutorial
refers to /var/www/html, it's also necesary to add aditional steps
to return the configuration back to the original values to leave
it ready for the user to be able to follow the next steps in this
and the WordPress tutorials.

Signed-off-by: Jair Gonzalez <jair.de.jesus.gonzalez.plascencia@intel.com>