Added note to increase size of VM partition
I added a note that might confuse new users when following the getting started sequence download, import vmware VM->enable user space. The linux error will not indicate that the hard drive isn't big enough. The link did not render in the preview for some reason so edit it to conform with your relative links convention and build before pulling.