Fix for document inconsistencies
* Changed directory name for mixer workspace where the workspace is given
a path and further in the document the workspace path different, this
leads to an error when using correct procedure specified in document.
This change uses the same directory path.

* Fix for a typo where the correct default value for a configuration
file name is incorrect and when copying value from document will lead to
errors when following proper procedure specified in document.

Signed-off-by: Alex Jaramillo <alex.v.jaramillo@intel.com>