bootable-usb: Add extra caution about unmount
Not unmounting an USB partition before an image is burned could cause
file system checksum errors leading to other issues (including kernel
panics) when trying to boot from the USB drive. This commit adds an extra
caution note to warn the users about the possible issues they could find
when this step is not followed correctly.

Signed-off-by: Jair Gonzalez <jair.de.jesus.gonzalez.plascencia@intel.com>