Add examples to support DLRS v4 release (#738)
* Add examples to support DLRS v4 release

* Add AIXPRT and PyTorch training examples to DLRS guide

* adds necessary white space before code block

* updates openvino examples

* adds links to referenced scripts

* Move dlrs into tutorials for immediate publication

The dlrs guide and images are moving -- the images into the source/_figures directory, and the dlrs file into tutorials, so they can be immediately published. This change enables partners working with DLRS v4 to continue development.  When possible the dlrs.rst file will be returned to the guides section, and further edits added then.

Signed-off-by: Beth Dean <beth.e.dean@intel.com>